requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
email-validator>=1.1.3
strawberry-graphql[fastapi]>=0.138.0
motor>=3.3.0
faker>=8.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
